optimization

+++ shadowsocksr.sh
@@ -147,8 +147,13 @@ install_pkg(){
   setuptools_url=https://files.pythonhosted.org/packages/68/75/d1d7b7340b9eb6e0388bf95729e63c410b381eb71fe8875cdfd949d8f9ce/setuptools-45.2.0.zip
   file_name=$(basename $setuptools_url)
   dir_name=${file_name%.*}
-  wget -O $file_name $setuptools_url
-  unzip $file_name
+  if [[ ! -f $file_name ]]; then
+    wget -O $file_name $setuptools_url
+    unzip $file_name
+    # delete zip files
+    rm -rf $dir_name.zip
+    rm -rf shadowsocksr.sh
+  fi
   cd $dir_name
 
   #install setuptools
@@ -329,7 +334,7 @@ main(){
   #check root permission
   isRoot=$( isRoot )
   if [[ "${isRoot}" != "true" ]]; then
-    echo -e "${RED_COLOR}error:${NO_COLOR}Please run this script as as root"
+    echo "${RED_COLOR}error:${NO_COLOR}Please run this script as as root"
     exit 1
   fi
   install_pkg
